Extend Rename to support update in PUT repository/name

+++ repository/repository.go
@@ -13,6 +13,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"reflect"
 	"regexp"
 	"strings"
 
@@ -178,32 +179,57 @@ func Remove(name string) error {
 	return nil
 }
 
-// Rename renames a repository.
-func Rename(oldName, newName string) error {
-	log.Debugf("Renaming repository %q to %q", oldName, newName)
-	repo, err := Get(oldName)
+// Update update a repository data.
+func Update(name string, newData Repository) error {
+	log.Debugf("Updating repository %q data", name)
+	repo, err := Get(name)
 	if err != nil {
-		log.Errorf("repository.Rename: Repository %q not found: %s", oldName, err)
+		log.Errorf("repository.Update: Repository %q not found: %s", name, err)
 		return err
 	}
-	newRepo := repo
-	newRepo.Name = newName
 	conn, err := db.Conn()
 	if err != nil {
 		return err
 	}
 	defer conn.Close()
-	err = conn.Repository().Insert(newRepo)
-	if err != nil {
-		log.Errorf("repository.Rename: Error adding new repository %q: %s", newName, err)
-		return err
+	if len(newData.Name) > 0 && newData.Name != repo.Name {
+		oldName := repo.Name
+		log.Debugf("Renaming repository %q to %q", oldName, newData.Name)
+		err = conn.Repository().Insert(newData)
+		if err != nil {
+			log.Errorf("repository.Rename: Error adding new repository %q: %s", newData.Name, err)
+			return err
+		}
+		err = conn.Repository().RemoveId(oldName)
+		if err != nil {
+			log.Errorf("repository.Rename: Error removing old repository %q: %s", oldName, err)
+			return err
+		}
+		err = fs.Filesystem().Rename(barePath(oldName), barePath(newData.Name))
+		if err != nil {
+			log.Errorf("repository.Rename: Error renaming old repository in filesystem %q: %s", oldName, err)
+		}
+	} else {
+		d := bson.M{}
+		if !reflect.DeepEqual(newData.Users, repo.Users) {
+			d["users"] = newData.Users
+		}
+		if !reflect.DeepEqual(newData.ReadOnlyUsers, repo.ReadOnlyUsers) {
+			d["readonlyusers"] = newData.ReadOnlyUsers
+		}
+		if !reflect.DeepEqual(newData.IsPublic, repo.IsPublic) {
+			d["ispublic"] = newData.IsPublic
+		}
+		if len(d) == 0 {
+			return nil
+		}
+		err = conn.Repository().UpdateId(repo.Name, bson.M{"$set": d})
+		if err != nil {
+			log.Errorf("repository.Update: Error updating repository data %q: %s", repo.Name, err)
+			return err
+		}
 	}
-	err = conn.Repository().RemoveId(oldName)
-	if err != nil {
-		log.Errorf("repository.Rename: Error removing old repository %q: %s", oldName, err)
-		return err
-	}
-	return fs.Filesystem().Rename(barePath(oldName), barePath(newName))
+	return nil
 }
 
 // ReadWriteURL formats the git ssh url and return it. If no remote is configured in
@@ -283,22 +309,6 @@ func (r *Repository) isValid() (bool, error) {
 		return false, errors.New("Validation Error: repository should have at least one user")
 	}
 	return true, nil
-}
-
-// SetAccess gives full or read-only permission for users in all specified repositories.
-// It redefines all users permissions, replacing the respective user collection
-func SetAccess(rNames, uNames []string, readOnly bool) error {
-	conn, err := db.Conn()
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-	if readOnly {
-		_, err = conn.Repository().UpdateAll(bson.M{"_id": bson.M{"$in": rNames}}, bson.M{"$set": bson.M{"readonlyusers": uNames}})
-	} else {
-		_, err = conn.Repository().UpdateAll(bson.M{"_id": bson.M{"$in": rNames}}, bson.M{"$set": bson.M{"users": uNames}})
-	}
-	return err
 }
 
 // GrantAccess gives full or read-only permission for users in all specified repositories.
